backend/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
elevenlabs==0.2.27
python-dotenv==1.0.0
requests>=2.28.0
# Gemini & File Processing
google-generativeai>=0.3.0
PyPDF2>=3.0.0
python-pptx>=0.6.23
Pillow>=10.0.0
# CORS for frontend
flask-cors>=4.0.0